refactor(ui): split search_results_list + profile_edit_sheet via part directive
- search_results_list.dart: 475 → 264 LOC. Static helpers
  (_computeCheapestFlagsFor, _getPriceRangeFor, _sortSearchResults,
  _fuelStationsFrom) and private widgets (_CollapsibleBrandFilters,
  _ViewToggleButton) moved to search_results_list_parts.dart.
- profile_edit_sheet.dart: 452 → 177 LOC. All six section widgets
  (_RouteSegmentSection, _TogglesSection, _RatingModeSection,
  _CountrySection, _LanguageSection, _SaveDeleteActions) moved to
  profile_edit_sheet_parts.dart.
- profile_edit_sheet_test.dart: source-level regression now reads
  both main + part file since the wiring it checks spans the library.

No public API change, no behavior change. All 3734 tests pass.
Progress on #563.
